Rename `search` -> `findLineIndex`
