+++ src/utils/collapsedRow/index.ts
@@ -1,10 +1,10 @@
 import { ProcessedLogLine } from "types/logs";
 
 /**
- * isCollapsedRow determines if a row is a collapsed row
- * although it is a simple function its purpose is to make the code more readable
+ * `isCollapsedRow` determines if a row is a collapsed row. Although it is a simple function, its purpose
+ * is to make the code more readable.
  * @param logLine ProcessedLogLine
- * @returns boolean true if the row is a collapsed row
+ * @returns {boolean} true if the row is a collapsed row
  */
 const isCollapsedRow = (logLine: ProcessedLogLine): logLine is number[] =>
   Array.isArray(logLine);
