+++ src/utils/findLineIndex/findLineIndex.test.ts
